feat: update user group check
