+++ RyoURL/shortURL/apis/__init__.py
@@ -1,3 +1,5 @@
+from .short_url_basic_api import short_url_router
+from .short_url_with_auth_api import auth_short_url_router
 from .auth_api import auth_router
-from .short_url_api import url_router
+from .user_api import user_router
 from .admin_api import admin_router
