+++ RyoURL/shortURL/apis/user_api.py
@@ -1,18 +1,23 @@
 from ninja import Router
 from rest_framework_simplejwt.tokens import RefreshToken
 from rest_framework_simplejwt.exceptions import TokenError
+from ninja.errors import HttpError
 
 from .schemas import UserResponseSchema, ErrorSchema, TokenSchema, TokenResponseSchema
 from ..models import User
-from .auth_api import JWTAuth
+from .auth import JWTAuth
 
-user_router = Router(tags=["user"])
+user_auth = JWTAuth()
 
-@user_router.get("info", auth=JWTAuth(), response={200: UserResponseSchema, 403: ErrorSchema, 404: ErrorSchema})
+class UserRouter(Router):
+    def get_permissions(self, request):
+        if not user_auth.user_check(user_auth.authenticate(request, request.headers.get('Authorization', '').split(' ')[-1])):
+            raise HttpError(403, "需要登入")
+
+user_router = UserRouter(tags=["user"])
+
+@user_router.get("info", auth=user_auth, response={200: UserResponseSchema, 403: ErrorSchema, 404: ErrorSchema})
 def get_user_info(request, username: str):
-    if not request.auth:
-        return 403, {"message": "需要登入"}
-    
     if username == request.auth['user'].username or request.auth['user_type'] == 2:
         try:
             user = User.objects.get(username=username)
@@ -22,7 +27,7 @@ def get_user_info(request, username: str):
     else:
         return 403, {"message": "無權限查看其他使用者資訊"}
 
-@user_router.post("refresh-token", auth=JWTAuth(), response={200: TokenResponseSchema, 400: ErrorSchema})
+@user_router.post("refresh-token", auth=user_auth, response={200: TokenResponseSchema, 400: ErrorSchema})
 def refresh_token(request, token_data: TokenSchema):
     try:
         refresh = RefreshToken(token_data.refresh)
